restructure account directory for anon and user layouts
