+++ src/app/account/(anon)/keys/page.tsx
@@ -1,15 +1,10 @@
-import { GenerateKeys } from './generate-keys'
 import { getReviewerPublicKeyAction } from '@/server/actions/user-keys.actions'
-import { RegenerateKeys } from './regenerate-keys'
+import { GenerateKeys } from './generate-keys'
 
 export const dynamic = 'force-dynamic'
 
 export default async function KeysPage() {
     const publicKey = await getReviewerPublicKeyAction()
 
-    if (publicKey) {
-        return <RegenerateKeys />
-    } else {
-        return <GenerateKeys />
-    }
+    return <GenerateKeys isRegenerating={!!publicKey} />
 }
