+++ src/app/account/(anon)/reset-password/pending-reset.tsx
