+++ src/app/account/(auth)/manage-key/page.tsx
@@ -1,0 +1,7 @@
+import { RegenerateKeys } from './regenerate-keys'
+
+export const dynamic = 'force-dynamic'
+
+export default async function ManageKeysPage() {
+    return <RegenerateKeys />
+}
