+++ src/app/account/layout.tsx
@@ -1,6 +1,5 @@
-import { AnonLayout } from '@/components/layout/anon-layout'
 import { type ReactNode } from 'react'
 
 export default async function AccountLayout({ children }: Readonly<{ children: ReactNode }>) {
-    return <AnonLayout>{children}</AnonLayout>
+    return children
 }
